requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
Django
django-cors-headers
django-filter
django-jazzmin
django-phonenumber-field
django-rest-framework
django-rest-swagger
djangorestframework
djangorestframework-simplejwt
idna
itypes
Jinja2
MarkupSafe
openapi-codec
phonenumbers
phonenumberslite
Pillow
PyJWT
pytz
requests
simplejson
sorl-thumbnail
sqlparse
twilio
uritemplate
urllib3
